+++ nexus_constructor/geometry_loader.py
@@ -1,21 +1,7 @@
 from nexus_constructor.data_model import OFFGeometry, Vector
 from nexusutils.readwriteoff import parse_off_file
+from nexus_constructor.unit_converter import calculate_unit_conversion_factor
 from stl import mesh
-import pint
-
-
-def calculate_unit_conversion_factor(units):
-    """
-    Determines the factor for multiplying the geometry file points in order to convert it from its original units to
-    metres.
-
-    :param units: A unit of length in the form of a string.
-    :return: A float value for converting between metres and the unit argument.
-    """
-    ureg = pint.UnitRegistry()
-    units = ureg(units)
-    units = ureg.m.from_(units)
-    return units.magnitude
 
 
 def load_geometry(filename: str, units: str, geometry: OFFGeometry=OFFGeometry()):
