+++ resources/Qt models/ValidUnits.qml
@@ -4,6 +4,6 @@ QtObject {
     // Booleans to indicate that the units provided for loading a geometry or creating a cylinder were valid
     property bool validMeshUnits: false
     property bool validCylinderUnits: false
-    property bool cylinderUnitMessage: false
+    property bool showCylinderUnitMessage: false
     readonly property string invalidUnitsText: "Units not recognised. Please enter a different type."
 }
