+++ tests/test_unit_conversion.py
@@ -1,0 +1,11 @@
+from nexus_constructor.unit_converter import calculate_unit_conversion_factor
+from pytest import approx
+
+
+def test_unit_conversion_factor():
+
+    # List of units and their expected value in metres
+    units = [("cm", 0.01), ("km", 1000), ("m", 1.0), ("inch", 0.0254), ("foot", 0.3048)]
+
+    for unit in units:
+        assert approx(calculate_unit_conversion_factor(unit[0])) == unit[1]
